:sparkle: standard linker scripts + renaming
- Rename to libhal-lpc40xx
- Add standard linker scripts to the linker/ directory
- Add standard linker scripts to package
- Change cmake target name to libhal::lpc40xx
- Rename include directory and libhal-lpc40xx

Resolves #119
