--- ARCHITECTURE.md
@@ -1,7 +1,0 @@
-# libhal Library Architecture
-
-## Library Organization
-
-The directory and file structure is based directly after the C++ std
-[Canonical Project Structure](http://www.open-std.org/jtc1/sc22/wg21/docs/papers/2018/p1204r0.html)
-proposal.
