+++ demos/applications/can.cpp
@@ -2,11 +2,11 @@
 #include <array>
 #include <string_view>
 
-#include <libarmcortex/dwt_counter.hpp>
+#include <libhal-armcortex/dwt_counter.hpp>
+#include <libhal-lpc40xx/can.hpp>
+#include <libhal-lpc40xx/uart.hpp>
 #include <libhal-util/serial.hpp>
 #include <libhal-util/steady_clock.hpp>
-#include <liblpc40xx/can.hpp>
-#include <liblpc40xx/uart.hpp>
 
 [[nodiscard]] hal::status application()
 {
