+++ demos/conanfile.txt
@@ -1,7 +1,10 @@
 [requires]
-liblpc40xx/0.3.2
+libhal-lpc40xx/0.3.4
+
+[tool_requires]
+gnu-arm-embedded-toolchain/11.3.0
 
 [generators]
 CMakeToolchain
 CMakeDeps
-VirtualRunEnv
+VirtualBuildEnv
