+++ include/libhal-lpc40xx/LICENSE
