+++ include/libhal-lpc40xx/adc.hpp
