+++ include/libhal-lpc40xx/internal/platform_check.hpp
