+++ linkers/libhal-lpc40xx/lpc4072.ld
@@ -4,4 +4,4 @@ __ram = 0x10000000;
 __ram_size = 16K;
 __stack_size = 1K;
 
-INCLUDE "common.ld"
+INCLUDE "libhal-armcortex/standard.ld"
