+++ linkers/libhal-lpc40xx/lpc4076.ld
@@ -1,7 +1,7 @@
 __flash = 0x00000000;
-__flash_size = 512K;
+__flash_size = 256K;
 __ram = 0x10000000;
 __ram_size = 64K;
 __stack_size = 1K;
 
-INCLUDE "common.ld"
+INCLUDE "libhal-armcortex/standard.ld"
