+++ linkers/libhal-lpc40xx/lpc4088.ld
@@ -4,4 +4,4 @@ __ram = 0x10000000;
 __ram_size = 64K;
 __stack_size = 1K;
 
-INCLUDE common.ld
+INCLUDE "libhal-armcortex/standard.ld"
