+++ test_package/main.cpp
@@ -1,5 +1,5 @@
 #include <cstdio>
-#include <liblpc40xx/input_pin.hpp>
+#include <libhal-lpc40xx/input_pin.hpp>
 
 int main()
 {
