+++ tests/conanfile.txt
@@ -1,7 +1,7 @@
 [requires]
 boost-ext-ut/1.1.9
 libhal/0.3.2
-libarmcortex/0.3.3
+libhal-armcortex/0.3.5
 libhal-util/0.3.3
 ring-span-lite/0.6.0
