+++ tests/system_controller.test.cpp
@@ -1,1 +1,1 @@
-#include <liblpc40xx/system_controller.hpp>
+#include <libhal-lpc40xx/system_controller.hpp>
